Fix app hang on exit if an exception occurs
If an exception occurs in any of the threads of the app, the user will
see that as an app hang.  The reason it appears to hang (rather than
exiting after the fatal uncaught exception occurs) is that the logging
subsystem wasn't marking its thread as a daemon thread.

Since there was a remaining !daemon thread still running python doesn't
exit.  The correct setting for a background thread like that mp_logging
thread is daemon=true.