Fix version patching when version comment contains a slash.
The release script failed when tag was of the following form:

git tag -a v3.2.0.dev0 -m "WIP by @geeksville, testing/using release binaries"

The root cause was that the sed expansion was using / as the replacement delimter (which appeared in the tag message).  This breaks sed replacement.

The 'fix' is to change the sed invoke to use a rare symbol instead of a common one.  / -> ^