add workaround for macos CI runner bug
The currently available github macos runners have a bug that cause occasional create-dmg failures
with the message "hdiutil: create failed - Resource busy"

In my testing it seems to occur about 50% of the time.

See this issue for (lots) of discussion.  There is currently no alternative to this fix if you want to build macos images on github.

https://github.com/actions/runner-images/issues/7522#issuecomment-2534682126