Fix MacOS CI build to work on 10.15
The create-dmg script, by default, tries to use an old hdiutil feature that Apple removed for security reasons in macOS 10.15 (Catalina).

> Apple deprecated and then completely removed this feature in macOS 10.15 Catalina because it posed a significant security risk. Automatically mounting and running content downloaded from the web is a great way to get malware.

By disabling this feature, Apple ensured that users must manually open the .dmg and drag the application to install it, which is a deliberate action that confirms user intent.